Support for encoded urls
Until that time, if for an argument -v was specified a url in which there is an ampersand, then the program started running in the background and it was almost impossible to stop it without re-opening the console or rebooting the whole system. This micro-patch fixes the issue. Now you should provide urlencoded link if there's an ampersand in it.